requirements.txt
flask
flask-restful
flask-sqlalchemy
flask-migrate
flask-script
pymysql
flask-jwt-extended
requests